requirements.txt
websocket
https://github.com/s4w3d0ff/python-poloniex/archive/v0.4.6.zip
service_identity
autobahn[twisted]
python-dateutil
